[lld-macho] Pre-layout exports trie
The exports trie `updateOffsets` was doing repeated work calculating the
node size in every loop until the fixed point, and in the first loop, it
was using `offset = 0` for every child, which under-sized the node,
forcing a second loop (only tries that do not have children beyond 127
offset will be correct, which is not likely in real world scenarios).

The idea is pre-calculating an estimation of the size of the node that
doesn't change between loops to the fixed point (terminal size, number
of children, length of the children substrings), and calculating the
minimal offset the next nodes will need to be (which depends on the
offset of the current node).

In theory this should have been faster, and it might be faster for just
the given layout code, but overall, the layout seems not to be as
critical as other parts, and whatever improvements it might bring, it is
not enough with the "noise" introduced by other parts of the linker.

Abandoning the approach, since it is more complicated, and requires
slightly more memory per node for no apparent benefit.